seperate logo component
