+++ Logo.js
@@ -1,0 +1,28 @@
+import React from 'react';
+import {View, Image, StyleSheet } from 'react-native';
+
+const Logo = () => {
+  return (
+    <View style={styles.container}>
+      <Image 
+       source={require('./image/logo.png')} 
+        style={styles.logo}
+      />
+    </View>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: "center",
+    alignItems: "center",
+  },
+  logo: {
+    width: 225,
+    height: 225,
+    resizeMode: "contain",
+  },
+});
+
+export default Logo;
